Moved `InjectDatabaseStateIntoException` trait
