+++ tests/Unit/InjectDatabaseStateIntoExceptionTest.php
@@ -8,11 +8,12 @@
 
 namespace Tests\Unit;
 
+use App\Traits\Tests\InjectDatabaseStateIntoException;
 use Tests\TestCase;
 
 class InjectDatabaseStateIntoExceptionTest extends TestCase {
 
-    // NOTE: InjectDatabaseStateIntoException trait has already been included by the TestCase class
+    use InjectDatabaseStateIntoException;
 
     public function testSettingValidInjectionPermission(){
         $current_state = $this->isDatabaseStateInjectionAllowed();
